fix(test): split protected-branch tests to satisfy 800-line gate (#1466)
- Move 13 protected-branch tests to test_pr_post_merge_check_protected.py
- Original test file: 718 lines (was 876)
- New test file: 266 lines
- No test semantics changed, no tests removed
- Minimal helper duplication between files
