+++ tests/unit/test_pr_post_merge_check.py
@@ -715,162 +715,3 @@ def test_ci_check_failed_fails():
     result = pp._check_ci(ci_data)
     assert len(result) >= 1
     assert any("failure" in r for r in result)
-
-
-# ---------------------------------------------------------------------------
-# Protected branch checks
-# ---------------------------------------------------------------------------
-
-
-def _make_all_pass_mocks():
-    """Return (pr_mock, ci_mock, git_mock) for tests where checks would pass."""
-    pr_mock = mock.MagicMock()
-    pr_mock.stdout = json.dumps(MERGED_PR_JSON)
-    pr_mock.returncode = 0
-
-    ci_mock = mock.MagicMock()
-    ci_mock.stdout = json.dumps(VALID_CI_RUN)
-    ci_mock.returncode = 0
-
-    git_mock = _make_git_mock()
-
-    return pr_mock, ci_mock, git_mock
-
-
-def _make_combined_run(pr_mock, ci_mock, git_mock):
-    """Return a combined subprocess.run mock."""
-
-    def _run(args, **kwargs):
-        if args[0] == "gh":
-            if "pr" in args:
-                return pr_mock
-            return ci_mock
-        if args[0] == "git":
-            return git_mock(args, **kwargs)
-        raise RuntimeError(f"Unexpected command: {args}")
-
-    return _run
-
-
-def test_protected_branch_main_fails():
-    """CONFIRM_CLEANUP=1 + BRANCH=main → FAIL, no deletion."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-    combined = _make_combined_run(pr_mock, ci_mock, git_mock)
-
-    with mock.patch("subprocess.run", side_effect=combined):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, "main")
-    _assert_fail(result, reason_contains="protected")
-    assert any("main" in f for f in result.failures)
-
-
-def test_protected_branch_master_fails():
-    """CONFIRM_CLEANUP=1 + BRANCH=master → FAIL, no deletion."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-    combined = _make_combined_run(pr_mock, ci_mock, git_mock)
-
-    with mock.patch("subprocess.run", side_effect=combined):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, "master")
-    _assert_fail(result, reason_contains="protected")
-
-
-def test_protected_branch_origin_main_fails():
-    """CONFIRM_CLEANUP=1 + BRANCH=origin/main → FAIL, no deletion."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-    combined = _make_combined_run(pr_mock, ci_mock, git_mock)
-
-    with mock.patch("subprocess.run", side_effect=combined):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, "origin/main")
-    _assert_fail(result, reason_contains="protected")
-
-
-def test_protected_branch_origin_master_fails():
-    """CONFIRM_CLEANUP=1 + BRANCH=origin/master → FAIL, no deletion."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-    combined = _make_combined_run(pr_mock, ci_mock, git_mock)
-
-    with mock.patch("subprocess.run", side_effect=combined):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, "origin/master")
-    _assert_fail(result, reason_contains="protected")
-
-
-def test_current_branch_fails():
-    """CONFIRM_CLEANUP=1 + BRANCH=current branch → FAIL, no deletion."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-
-    # Make git rev-parse --abbrev-ref HEAD return the branch name
-    def _git_with_current(args, **kwargs):
-        if args[1] == "rev-parse" and "--abbrev-ref" in args:
-            result = mock.MagicMock()
-            result.returncode = 0
-            result.stdout = _BRANCH
-            result.stderr = ""
-            return result
-        return git_mock(args, **kwargs)
-
-    def _combined_run(args, **kwargs):
-        if args[0] == "gh":
-            if "pr" in args:
-                return pr_mock
-            return ci_mock
-        if args[0] == "git":
-            return _git_with_current(args, **kwargs)
-        raise RuntimeError(f"Unexpected command: {args}")
-
-    with mock.patch("subprocess.run", side_effect=_combined_run):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, _BRANCH)
-    _assert_fail(result, reason_contains="currently checked-out")
-
-
-def test_cleanup_branch_refuses_protected():
-    """cleanup_branch() refuses main even if called directly."""
-    logs = pp.cleanup_branch("main")
-    assert any("REFUSED" in line for line in logs)
-
-
-def test_cleanup_branch_refuses_master():
-    """cleanup_branch() refuses master even if called directly."""
-    logs = pp.cleanup_branch("master")
-    assert any("REFUSED" in line for line in logs)
-
-
-def test_cleanup_branch_refuses_origin_main():
-    """cleanup_branch() refuses origin/main even if called directly."""
-    logs = pp.cleanup_branch("origin/main")
-    assert any("REFUSED" in line for line in logs)
-
-
-def test_cleanup_branch_refuses_origin_master():
-    """cleanup_branch() refuses origin/master even if called directly."""
-    logs = pp.cleanup_branch("origin/master")
-    assert any("REFUSED" in line for line in logs)
-
-
-def test_check_branch_not_protected_ok():
-    """Normal feature branch passes protection check."""
-    result = pp._check_branch_not_protected(_BRANCH)
-    assert len(result) == 0
-
-
-def test_check_branch_not_protected_main():
-    """main is detected as protected."""
-    result = pp._check_branch_not_protected("main")
-    assert len(result) >= 1
-    assert any("protected" in r.lower() for r in result)
-
-
-def test_check_branch_not_protected_master():
-    """master is detected as protected."""
-    result = pp._check_branch_not_protected("master")
-    assert len(result) >= 1
-    assert any("protected" in r.lower() for r in result)
-
-
-def test_non_protected_branch_cleanup_allowed():
-    """Normal branch with CONFIRM_CLEANUP can proceed through evaluate."""
-    pr_mock, ci_mock, git_mock = _make_all_pass_mocks()
-    combined = _make_combined_run(pr_mock, ci_mock, git_mock)
-
-    # _BRANCH is "feature/test-branch" — not in PROTECTED_BRANCHES
-    with mock.patch("subprocess.run", side_effect=combined):
-        result = pp.evaluate(_PR_NUM, _MERGE_COMMIT, _BRANCH)
-    _assert_pass(result)
